add uart control and machine number

+++ uart.h
@@ -7,7 +7,7 @@ extern void UART_Init(void);
 extern void UART_Update(void);
 extern void UART_PutChar (char c);
 extern void UART_TimeTicks(void);
-extern void UART_Send(u8 * reply, u8 count);
+extern void UART_Send(u8 * reply, u16 count);
 
 @far @interrupt void UART_ISR(void);
